Add list functionality and related tests to InternxtClient
This commit introduces the ability to list the items in the root or specified folder in the InternxtClient. To verify this new functionality, corresponding tests have been added in InternxtClientListTests. Additionally, a new model, InternxtItem, has been added to represent listed items, and a base class for test setup, InternxtClientTestBase, has also been introduced.