Add upload functionality and related tests
This commit incorporates the ability to upload files in the Internxt SDK. It introduces two new classes, the ResultParser and the InternxtUploadResult. The ResultParser is employed to parse data returned from the API, and tests have been added under the InternxtClientUploadTests class to ensure file upload functions as expected. In addition, the test file 'test.txt' is included in the project for testing purposes.