Update dotnet-desktop workflow in GitHub Actions
The dotnet-desktop workflow configuration for GitHub Actions has been updated to run on branches from both push and pull request triggers. Furthermore, the publish and push to NuGet step is now restricted to when the GitHub reference starts with 'refs/tags/'. These changes aim for a more versatile use of the CI pipeline, especially related to development branches, while also maintaining the security of publishing packages only when a new tag is created.