Update action for publishing to NuGet
The changes refactored the GitHub action to publish to NuGet. Specifically, it is now triggered by tags starting with 'v' and the publish step was moved to its own job, which happens after a successful build. In addition, the dotnet pack and nuget push commands were updated to better handle the package versioning using the tag name.